+++ src/common.cc
@@ -166,6 +166,8 @@ namespace sharp {
     descriptor->access = AttrAsBool(input, "sequentialRead") ? VIPS_ACCESS_SEQUENTIAL : VIPS_ACCESS_RANDOM;
     // Remove safety features and allow unlimited input
     descriptor->unlimited = AttrAsBool(input, "unlimited");
+    // Use the EXIF orientation to auto orient the image
+    descriptor->autoOrient = AttrAsBool(input, "autoOrient");
     return descriptor;
   }